+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.4.0
+## Features
+* Package requirements into service root directory in order to avoid munging
+  sys.path to load requirements (#30).
+* Package requirements when deploying individual non-WSGI functions (#30).
+* Added `pythonBin` option to set python executable, defaulting to current runtime version (#29).
+
+
 # 1.3.1
 ## Features
 * Add configuration for handling base path mappings (API_GATEWAY_BASE_PATH)